+++ lib/src/models/models.dart
@@ -6,6 +6,7 @@ export './device_info.dart';
 export './event.dart';
 export './exception.dart';
 export './faro_user.dart';
+export './installation_id.dart';
 export './log.dart';
 export './log_level.dart';
 export './measurement.dart';